Add dedicated db directory
This makes the directory structure less messy.
It also makes it cleaner to ignore the json in the `.gitignore`.

Data should not be stored in the repo because it will persist for user.

Path to json data should include db subdirectory.

Git must ignore any json files created in models/db.